requirements.txt
Django==1.7.1
Fabric==1.8.2
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.1
PyYAML==3.11
Whoosh==2.6.0
ansible==1.7.2
argparse==1.2.1
boto==2.34.0
django-boto==0.3.8
django-debug-toolbar==1.2.1
django-haystack==2.3.1
django-storages==1.1.8
ecdsa==0.10
gunicorn==19.1.1
mock==1.0.1
paramiko==1.12.2
psycopg2==2.5.4
pycrypto==2.6.1
requests==2.2.1
selenium==2.44.0
simplejson==3.3.3
sqlparse==0.1.13
stevedore==0.15
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3
wsgiref==0.1.2
